* origin/4.x-up: (53 commits)
  order typeaheads by label by default
  Add some text filters for twig.
  more tests
  do not remove files in test environmenet
  Fix some tests.
  fixtures and some bug fixes
  More documentation
  fix name
  more docs
  documentation
  more test generators
  make sure reset tokens work once.
  bump some composer dependency versions
  reexamine the permissions
  refactor the comment tests so as not to couple too closely to blog stuff
  Add a phpunit extension to print a message for tests that take too long
  add more terrible urls to test the referer stuff
  rethink the way successful logins redirect
  shorten some indexed string fields to fit the index limitiation in mariadb
  allow spaces in image file names
  ...

+++ DublinCoreBundle/Service/ValueManager.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 /*
- * (c) 2021 Michael Joyce <mjoyce@sfu.ca>
+ * (c) 2022 Michael Joyce <mjoyce@sfu.ca>
  * This source file is subject to the GPL v2, bundled
  * with this source code in the file LICENSE.
  */
@@ -13,46 +13,31 @@ namespace Nines\DublinCoreBundle\Service;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\EventSubscriber;
 use Doctrine\Common\Util\ClassUtils;
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Doctrine\ORM\Events;
+use Exception;
 use Nines\DublinCoreBundle\Entity\Value;
 use Nines\DublinCoreBundle\Entity\ValueInterface;
 use Nines\DublinCoreBundle\Repository\ValueRepository;
-use Nines\MediaBundle\Entity\Link;
-use Nines\MediaBundle\Service\AbstractFileManager;
 use Nines\UtilBundle\Entity\AbstractEntity;
 
 /**
  * Link management service for Symfony.
  */
-class ValueManager extends AbstractFileManager implements EventSubscriber {
-    /**
-     * @var ValueRepository
-     */
-    private $valueRepository;
+class ValueManager implements EventSubscriber {
+    private ?ValueRepository $valueRepository = null;
 
-    /**
-     * @required
-     */
-    public function setValueRepository(ValueRepository $valueRepository) : void {
-        $this->valueRepository = $valueRepository;
-    }
-
-    /**
-     * Check if an entity is configured to accept values.
-     */
-    public function acceptsValues(AbstractEntity $entity) : bool {
-        return $entity instanceof ValueInterface;
-    }
+    private ?EntityManagerInterface $em = null;
 
     /**
      * Find the values for an entity.
      *
-     * @param mixed $entity
+     * @param AbstractEntity|ValueInterface $entity
      *
-     * @return Collection|Value[]
+     * @return Value[]
      */
-    public function findValues($entity) {
+    public function findValues($entity) : array {
         $class = ClassUtils::getClass($entity);
 
         return $this->valueRepository->findBy([
@@ -60,7 +45,15 @@ class ValueManager extends AbstractFileManager implements EventSubscriber {
         ]);
     }
 
-    public function setValues(ValueInterface $entity, Collection $values) : void {
+    /**
+     * @param Collection<int,Value>|Value[] $values
+     *
+     * @throws Exception
+     */
+    public function setValues(AbstractEntity $entity, $values) : void {
+        if ( ! $entity instanceof ValueInterface) {
+            throw new Exception(get_class($entity) . ' does not implement ValueInterface.');
+        }
         foreach ($this->findValues($entity) as $value) {
             $this->em->remove($value);
         }
@@ -70,7 +63,7 @@ class ValueManager extends AbstractFileManager implements EventSubscriber {
         }
     }
 
-    public function getSubscribedEvents() {
+    public function getSubscribedEvents() : array {
         return [
             Events::postLoad,
             Events::preRemove,
@@ -94,5 +87,23 @@ class ValueManager extends AbstractFileManager implements EventSubscriber {
         foreach ($entity->getValues() as $value) {
             $this->em->remove($value);
         }
+    }
+
+    /**
+     * @required
+     *
+     * @codeCoverageIgnore
+     */
+    public function setEntityManager(EntityManagerInterface $em) : void {
+        $this->em = $em;
+    }
+
+    /**
+     * @required
+     *
+     * @codeCoverageIgnore
+     */
+    public function setValueRepository(ValueRepository $valueRepository) : void {
+        $this->valueRepository = $valueRepository;
     }
 }
